+++ packages/blog/gatsby-blog-core/src/utils/default.options.js
@@ -39,6 +39,9 @@ module.exports = pluginOptions => {
       name: 'strapi',
       enabled: getValue(pluginOptions, 'sources.strapi', false),
       sourcePlugin: 'gatsby-source-strapi',
+      sourcePluginVersion: '2.0.0',
+      sourcePluginUpgradeDocs:
+        'https://elegantstack.netlify.app/flexiblog/sourcing-data/strapi-cms/#migrating-from-strapi-v3-to-v4',
       imageNodeType: 'ImageSharp',
       typeDefs: typesDefs.strapi
     }
@@ -124,6 +127,9 @@ module.exports = pluginOptions => {
   )
   const imageQuality = getValue(pluginOptions, 'imageQuality', 75)
 
+  const gatsbyRemarkPlugins = getValue(pluginOptions, 'gatsbyRemarkPlugins', [])
+  const remarkPlugins = getValue(pluginOptions, 'remarkPlugins', [])
+
   const pageContextOptions = {
     paginatePostsPage,
     basePath,
@@ -150,6 +156,8 @@ module.exports = pluginOptions => {
     paginatePostsPage,
     collectionPostsPerPage,
     slugSanitizeRegex,
-    pageContextOptions
+    pageContextOptions,
+    gatsbyRemarkPlugins,
+    remarkPlugins
   }
 }