+++ packages/flow-ui/flow-ui-theme/src/theme/global.js
@@ -1,6 +1,7 @@
 export default {
   html: {
-    fontSize: [`85%`, `90%`, `95%`, `100%`]
+    fontSize: [`85%`, `90%`, `95%`, `100%`],
+    scrollBehavior: `smooth`
   },
   '::selection': {
     color: t => t.colors.white,