+++ site/gatsby-config.js
@@ -1,55 +1,25 @@
 module.exports = {
   plugins: [
     {
-      resolve: 'gatsby-plugin-mailchimp',
+      resolve: '@elegantstack/gatsby-theme-flexiblog-agency',
       options: {
-        endpoint: "https://uylas.us7.list-manage.com/subscribe/post?u=56f18106e942eac04fac11bc7&amp;id=b37cfb1ab8"
-      }
-    },
-    {
-      resolve: "gatsby-plugin-google-analytics",
-      options: {
-        // The property ID; the tracking code won't be generated without it
-        trackingId: "UA-190518336-1"
-      },
-    },
-    {
-      resolve: "gatsby-plugin-disqus",
-      options: {
-        shortname: 'ibrahim-uylas',
-      },
-    },
-    {
-      resolve: "gatsby-plugin-sitemap",
-      options: {},
-    },
-    {
-      resolve: "gatsby-plugin-netlify-cms",
-      options: {},
-    },
-    {
-      // ATTENTION: Match the theme name with the theme you're using
-      resolve: "@elegantstack/gatsby-theme-flexiblog-agency",
-      options: {
-        services: {
-          mailchimp: true,
-          disqus: true,
-        },
+        // Add theme options here. Check documentation for available options.
+        siteUrl: process.env.URL || process.env.VERCEL_URL,
         sources: {
           local: true,
-        },
-      },
-    },
+        }
+      }
+    }
   ],
   // Customize your site metadata:
   siteMetadata: {
     siteUrl: "https://www.ibrahimuylas.com", // Your site URL without trailing slash,
     //General Site Metadata
     title: 'İbrahim Uylaş',
-     name: 'IbrahimUylas',
-     description: 'Kampçılık ve doğa yürüyüşleriyle ilgili ne ararsan var, ne vereyim abime!',
-     address: 'An itibariyle Londra',
-     email: 'ibrahim@uylas.net',
+    name: 'IbrahimUylas',
+    description: 'Kampçılık ve doğa yürüyüşleriyle ilgili ne ararsan var, ne vereyim abime!',
+    address: 'An itibariyle Londra',
+    email: 'ibrahim@uylas.net',
     phone: '',
 
     //Site Social Media Links