+++ packages/flow-ui/flow-ui-widgets/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/flow-ui-widgets@3.9.0...@elegantstack/flow-ui-widgets@4.0.0) (2023-03-11)
+
+
+### Features
+
+* upgrade to Gatsby v5 and update other dependencies ([aba9eb2](https://github.com/ElegantStack/gatsby-themes/commit/aba9eb243ebe39f0f02885b9693703a316c0a91f))
+
+
+### Bug Fixes
+
+* add React Suspense to fix hydration error ([95a16bc](https://github.com/ElegantStack/gatsby-themes/commit/95a16bcfde3839a957006cd57c91c46229deda06))
+* fix react warnings and hydration errors ([cd0f1e9](https://github.com/ElegantStack/gatsby-themes/commit/cd0f1e9dc127a426c95550e46b925047f1e7d1ca))
+
+
+### Refactor
+
+* **flow-ui-widgets:** improve Twitter sharing card meta tags ([837cce1](https://github.com/ElegantStack/gatsby-themes/commit/837cce123ced9ff0fb9660a9a126bdab7791d0c6))
+* lock mdx plugin version ([38dcc87](https://github.com/ElegantStack/gatsby-themes/commit/38dcc876b24f70df414c82ca8fb59f91d8d74af8))
+* refactor code for Gatsby v5 support ([f3852fd](https://github.com/ElegantStack/gatsby-themes/commit/f3852fd3b880cb7b77271f358b092ce730797334))
+
+
+
+
+## [3.9.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/flow-ui-widgets@3.8.0...@elegantstack/flow-ui-widgets@3.9.0) (2022-05-18)
+
+
+### Features
+
+* Upgrade Gatsby and React and other depenedencies ([144ccc1](https://github.com/ElegantStack/gatsby-themes/commit/144ccc14a11f5f5cb5721d744922a3f164410f17))
+
+
+
+## [3.8.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/flow-ui-widgets@3.7.0...@elegantstack/flow-ui-widgets@3.8.0) (2022-02-28)
+
+
+### Features
+
+* upgrade Gatsby and other dependencies ([c42e47d](https://github.com/ElegantStack/gatsby-themes/commit/c42e47d5aea6364d7671bf5f75fbed7cee431c73))
+
+
+### Refactor
+
+* **gatsby-plugin-smooth-scrollbar:** remove unused package ([4b08b77](https://github.com/ElegantStack/gatsby-themes/commit/4b08b77f1e0516dd52d33a09597c907401521009))
+
+
+
+
+## [3.7.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/flow-ui-widgets@3.6.0...@elegantstack/flow-ui-widgets@3.7.0) (2021-11-15)
+
+
+### Features
+
+* upgrade to Gatsby v4 ([e481e4a](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/e481e4ab705d20c9d3daf3b2048f29eef308f420))
+
+
+
+
 ## [3.6.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/flow-ui-widgets@3.5.0...@elegantstack/flow-ui-widgets@3.6.0) (2021-09-07)