+++ packages/themes/gatsby-theme-flexiblog-beauty/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-theme-flexiblog-beauty@3.10.0...@elegantstack/gatsby-theme-flexiblog-beauty@4.0.0) (2023-03-11)
+
+**Note:** Version bump only for package @elegantstack/gatsby-theme-flexiblog-beauty
+
+
+
+
+
+
+## [3.10.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-theme-flexiblog-beauty@3.9.0...@elegantstack/gatsby-theme-flexiblog-beauty@3.10.0) (2022-05-18)
+
+
+### Features
+
+* Upgrade Gatsby and React and other depenedencies ([144ccc1](https://github.com/ElegantStack/gatsby-themes/commit/144ccc14a11f5f5cb5721d744922a3f164410f17))
+
+
+
+## [3.9.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-theme-flexiblog-beauty@3.8.0...@elegantstack/gatsby-theme-flexiblog-beauty@3.9.0) (2022-02-28)
+
+**Note:** Version bump only for package @elegantstack/gatsby-theme-flexiblog-beauty
+
+
+
+
+
+
+## [3.8.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-theme-flexiblog-beauty@3.7.0...@elegantstack/gatsby-theme-flexiblog-beauty@3.8.0) (2021-11-15)
+
+
+### Features
+
+* upgrade to Gatsby v4 ([e481e4a](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/e481e4ab705d20c9d3daf3b2048f29eef308f420))
+
+
+
+
 ## [3.7.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-theme-flexiblog-beauty@3.6.0...@elegantstack/gatsby-theme-flexiblog-beauty@3.7.0) (2021-09-07)