+++ starters/gatsby-starter-flexiblog-science/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-starter-flexiblog-science@3.10.0...@elegantstack/gatsby-starter-flexiblog-science@4.0.0) (2023-03-11)
+
+
+### Features
+
+* upgrade to Gatsby v5 and update other dependencies ([aba9eb2](https://github.com/ElegantStack/gatsby-themes/commit/aba9eb243ebe39f0f02885b9693703a316c0a91f))
+
+
+### Refactor
+
+* enable DEV_SSR in gatsby ([abf906d](https://github.com/ElegantStack/gatsby-themes/commit/abf906dfe01a534371b94421296134df133ffcfa))
+* refactor code for Gatsby v5 support ([f3852fd](https://github.com/ElegantStack/gatsby-themes/commit/f3852fd3b880cb7b77271f358b092ce730797334))
+* remove deprecated markdown syntax from demo content ([5dc172e](https://github.com/ElegantStack/gatsby-themes/commit/5dc172e6ab6fbf8077f0926dbea42e9e440cd5ec))
+
+
+
+
+## [3.10.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-starter-flexiblog-science@3.9.0...@elegantstack/gatsby-starter-flexiblog-science@3.10.0) (2022-05-18)
+
+
+### Features
+
+* Upgrade Gatsby and React and other depenedencies ([144ccc1](https://github.com/ElegantStack/gatsby-themes/commit/144ccc14a11f5f5cb5721d744922a3f164410f17))
+
+
+
+## [3.9.0](https://github.com/ElegantStack/gatsby-themes/compare/@elegantstack/gatsby-starter-flexiblog-science@3.8.0...@elegantstack/gatsby-starter-flexiblog-science@3.9.0) (2022-02-28)
+
+
+### Features
+
+* upgrade Gatsby and other dependencies ([c42e47d](https://github.com/ElegantStack/gatsby-themes/commit/c42e47d5aea6364d7671bf5f75fbed7cee431c73))
+
+
+
+
+## [3.8.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-starter-flexiblog-science@3.7.0...@elegantstack/gatsby-starter-flexiblog-science@3.8.0) (2021-11-15)
+
+
+### Features
+
+* upgrade to Gatsby v4 ([e481e4a](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/commit/e481e4ab705d20c9d3daf3b2048f29eef308f420))
+
+
+
+
 ## [3.7.0](https://gitlab.com/alimoosavi15/gatsby-theme-flexiblog/compare/@elegantstack/gatsby-starter-flexiblog-science@3.6.0...@elegantstack/gatsby-starter-flexiblog-science@3.7.0) (2021-09-07)